+++ widgets/views/entry.php
@@ -1,11 +1,12 @@
 <?php
 
+use humhub\modules\content\widgets\richtext\RichText;
+use humhub\modules\polls\models\Poll;
 use yii\helpers\Html;
-use humhub\widgets\RichText;
 
 humhub\modules\polls\assets\PollsAsset::register($this);
 
-/** @var $poll \humhub\modules\polls\models\Poll **/
+/** @var $poll Poll **/
 ?>
 
 <div data-poll="<?= $poll->id ?>" data-content-component="polls.Poll" data-content-key="<?= $poll->content->id ?>">
@@ -18,22 +19,22 @@ humhub\modules\polls\assets\PollsAsset::register($this);
         &nbsp;<span class="label label-success pull-right"><?= Yii::t('PollsModule.widgets_views_entry', 'Anonymous') ?></span>
     <?php endif; ?>
 
-    <?= Html::beginForm($contentContainer->createUrl('/polls/poll/answer', ['pollId' => $poll->id])); ?>
+    <?= Html::beginForm($contentContainer->createUrl('/polls/poll/answer', ['pollId' => $poll->id])) ?>
 
     <div data-ui-markdown>
-        <?= RichText::widget(['text' => $poll->question]); ?>
+        <?= RichText::output($poll->question) ?>
     </div>
 
     <br><br>
 
     <?php foreach ($poll->getViewAnswers() as $answer) : ?>
-        <?= $this->render('_answer', ['poll' => $poll, 'answer' => $answer, 'contentContainer' => $contentContainer]); ?>
+        <?= $this->render('_answer', ['poll' => $poll, 'answer' => $answer, 'contentContainer' => $contentContainer]) ?>
     <?php endforeach; ?>
 
     <?php if(!$poll->isShowResult()) : ?>
         <br>
         <div class="alert alert-default" style="margin:0">
-            <?= Yii::t('PollsModule.widgets_views_entry', '<strong>Note:</strong> The result is hidden until the poll is closed by a moderator.'); ?>
+            <?= Yii::t('PollsModule.widgets_views_entry', '<strong>Note:</strong> The result is hidden until the poll is closed by a moderator.') ?>
         </div>
     <?php endif; ?>